requirements.txt
# Python dependencies for AI Employee System

# File system monitoring
watchdog==4.0.0

# Gmail API
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Browser automation
playwright==1.41.0

# Utilities
python-dotenv==1.0.1
requests==2.31.0
